Minor design improvements
