+++ src/lib/components/Progress/ProgressLine.svelte
@@ -1,11 +1,13 @@
 <script lang="ts">
-  import { cn } from "$lib"
+  import { cn } from "$lib/index.js"
+  import type { AsChild } from "$lib/types.js"
   import type { HTMLAttributes } from "svelte/elements"
   import { tv } from "tailwind-variants"
-  import ctx from "./ctx"
+  import ctx from "./ctx.js"
 
-  type $$Props = HTMLAttributes<HTMLDivElement>
+  type $$Props = HTMLAttributes<HTMLDivElement> & AsChild
 
+  export let asChild: $$Props["asChild"] = false
   export { className as class }
 
   let className = ""
@@ -13,9 +15,12 @@
   const style = tv({
     base: cn`
       transition-all
+      absolute
+      top-0
+      left-0
       h-full
       bg-primary
-      rounded-lg
+      rounded-xl
     `,
   })
 
@@ -26,4 +31,10 @@
   $: percentage = ($value! * 100) / $max
 </script>
 
-<div class={cn(style.base, className)} style="width: {percentage}%" {...$$restProps} />
+{#if asChild}
+  <slot />
+{:else}
+  <div class={cn(style.base, className)} style="width: {percentage}%" {...$$restProps}>
+    <slot />
+  </div>
+{/if}
